% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEF_lethonen2004_spc_age.R
\name{BEF_Let2004_spc_age}
\alias{BEF_Let2004_spc_age}
\title{Biomass expansion factors according to Lethonen etal 2004}
\usage{
BEF_Let2004_spc_age(species = c("gran", "furu", "lauv")[1], age = 80)
}
\arguments{
\item{species}{one of {"spruce", "gran", "pine", "furu", "broadleave", "løv", "lauv", "bjørk"}}

\item{age}{age of the stand (numeric, years from breast height)}
}
\value{
a data.frame with biomass expansion factors for estimating dry mass of
tree parts like stem, branches etc, in tonne per m3 stem volume
(likely total stump to top stem volume) (også kalt skogskubikmeter)
}
\description{
Biomass expansion factors according to Lethonen etal 2004
}
\examples{
BEF_Let2004_spc_age(species = "pine", age = 70)
}
