% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hogst_kommune_ld}
\alias{hogst_kommune_ld}
\title{Hogststatistikk per kommune, sortiment, år, ... fra 2014}
\format{
tibble med om lag 20k obs og 13 variabler
\describe{
\item{FYLKENR }{Nr paa fylke (paa hogsttidspunkt); chr "01" "01" ...}
\item{FYLKENAVN}{; "Østfold" "Østfold" ...}
\item{KOMNR}{Nr paa kommune (på hogsttidspunkt); chr "0101" "0101" ...}
\item{KOMNAVN}{; "HALDEN" "HALDEN" ...}
\item{AVVIRKAAR}{årstall; num  2014 2014 2014 ...}
\item{SORTKODE}{Sortimentkode : chr "01" "01" "01" ...}
\item{SORTIMENT}{Sortimentnavn: chr "Sagtømmer" "Sagtømmer" ... }
\item{VIRKESGRP}{Virkesgruppe; chr "1-Gran" "1-Gran" "2-Furu" "3-Lauv" "5-Ved" ... }
\item{VIRKESKAT}{Virkeskategori; chr "1110" "1140" "1148" ...}
\item{KATEGORITEKST}{sortimentgruppe chr "Gran spesial" "Gran sagtømmer sams" }
\item{TOTALVOLUM}{m3; num 3440 14215 ...}
\item{TOTALVERDI}{kr; num 1599470 6898324 703000 ...}
\item{M3PRIS}{kr pr m3; num 465 485 263 398 3 ...}
}
}
\usage{
hogst_kommune_ld
}
\description{
Data fra Landbruksdirektoratets hogststatistikk
}
\keyword{datasets}
