% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kpi_t03014}
\alias{kpi_t03014}
\title{Konsumprisindeks}
\format{
tibble med om lag 41 obs og 2 variabler
\describe{
\item{år}{årstall; chr "1979" "1980" ...}
\item{kpi}{kpi ; num 25.9, 28.7,  ...}
}
}
\source{
\url{https://www.ssb.no/statbank/table/03014}
}
\usage{
kpi_t03014
}
\description{
Data fra ssb tabell 03014:
Konsumprisindeks, etter konsumgruppe (2015=100) 1979 - 2019
For konsumgrupp totalindeks
}
\keyword{datasets}
