% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t08801_ImpExp_Trevarer.R
\name{plot_rw_export}
\alias{plot_rw_export}
\title{Plot roundwood export volume and prices}
\usage{
plot_rw_export()
}
\value{
a list of four plots
}
\description{
Plot roundwood export volume and prices
}
\examples{
pts <- plot_rw_export()
pts[[1]]
}
