% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_regioninndeling.R
\name{regnavn.at.ref.yr}
\alias{regnavn.at.ref.yr}
\title{Region navn og region kode for gitt tidspunkt}
\usage{
regnavn.at.ref.yr(regionstat, ref.yr = lubridate::year(lubridate::now()))
}
\arguments{
\item{regionstat}{is the regional statistics, where each obs represent one year
and one region (fylke eller kommune)
The regionstat must have the variables
"region_kode" (a Norwegian region code)
"aar" (year of registration of the statistics)}

\item{ref.yr}{is the reference year to which geographic name and code is to be used}
}
\value{
tibble having the regional statistics including the regional
names and codes for the reference year in question
}
\description{
Denne funksjonen tar regionkoder og regionnavn fra en regional statistikk,
henter inn tabell som viser historiske endringer i regional inndeling av Norge
og gjør om til riktige koder og navn for et gitt referanseår (ref.yr)
Funksjonen fungerer for fylkesnivå inkludert landet ELLER for kommunenivå.
}
\examples{
regnavn.at.ref.yr(regionstat = t12750()) \%>\% glimpse()
}
