% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regref_fylke}
\alias{regref_fylke}
\title{Regioninndeling fylkesvis siden 1994}
\format{
En tibble med 20 rader (20 fylker i '94) og 8 variabler:
\describe{
\item{reg_code_199401}{Regionkode januar 1994; chr "01" "02"...}
\item{reg_name_199401}{Regionnavn januar 1994; chr "Oestfold" "Akershus" ..}
\item{reg_code_200607}{Regionkode juli 2006; chr "01" "02"...}
\item{reg_name_200607}{Regionnavn juli 2006; chr "Oestfold" "Akershus" ...}
\item{...code...}{fortsetter paa samme vis til 2020}
\item{...name...}{fortsetter paa samme vis til 2020}
}
}
\source{
\url{https://www.ssb.no/klass/klassifikasjoner/127/versjoner}
}
\usage{
regref_fylke
}
\description{
En tabell som viser endringer i kommunenes navn og nummer fra utgangspunkt
januar 1994 fram til nå. Tabellen kan benyttes sammen med
SSB regional statistikk (paa kommunenivaa) for aa koble observasjoner fra
tidligere aar til dagens kommunestruktur.
Fylkesinndelingen og endringer er hentet her https://www.ssb.no/klass/klassifikasjoner/104
}
\keyword{datasets}
