% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regref_fylke_l}
\alias{regref_fylke_l}
\title{Region referansetabell for fylker}
\format{
data frame med om lag 80 obs og 8 variabler:
\describe{
\item{ymfrom}{År og måned reg_code_from og reg_name_from gjelder fra; chr  "197201" "197201" "197201"  ...}
\item{yfrom}{årstallet fra ymfrom; chr "1972" "1972" "1972"  ...}
\item{reg_code_from}{Kode på regionenhet fra dato "ymfrom" ; chr "01" "02" "03" ...}
\item{reg_name_from}{Navn på regionenhet fra dato "ymfrom"; "Østfold" "Akershus" "Oslo" ...}
\item{ymto}{År og måned reg_code_to og reg_name_to gjelder fra: chr "202001" "202001" "202001" }
\item{yto}{År reg_code_to og reg_name_to gjelder fra: chr "2020" "2020" "2020" }
\item{reg_code_to}{Kode på regionenhet fra dato ymto:chr "30" "30" "03" ...}
\item{reg_name_to}{Navn på regionenhet fra dato ymto: chr  "Viken" "Viken" "Oslo"  ...}
}
}
\source{
\url{https://www.ssb.no/klass/klassifikasjoner/127/versjoner}
}
\usage{
regref_fylke_l
}
\description{
Tabellen gjør det mulig å slå opp hva et fylkesnavn på et gitt tidspunkt
tilsvarer i dagens fylkesstruktur.
}
\keyword{datasets}
