% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortimentparser.R
\name{sortimentparser}
\alias{sortimentparser}
\title{Sortimentparser}
\usage{
sortimentparser(tbbl)
}
\arguments{
\item{tbbl}{a tibble having the variable "sortiment"}
}
\value{
a tibble
}
\description{
Sortimentparser
}
\examples{
df <- tibble::tibble(sortiment = c("Gran spesialtømmer","Furu massevirke" ))
sortimentparser(df)
}
