% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortimentparser.R
\name{sortimentparser_vsop}
\alias{sortimentparser_vsop}
\title{Sortimentparser_vsop}
\usage{
sortimentparser_vsop(tbbl)
}
\arguments{
\item{tbbl}{a tibble having the variables  "virkesgrp" and "kategoritekst" and "virkeskat"}
}
\value{
a tibble
}
\description{
Sortimentparser_vsop
}
\examples{
sortimentparser_vsop(avvirk_kmn_ldir_regkorigert) \%>\% print(n = 20)
sortimentparser_vsop(avvirk_kmn_ldir_regkorigert) \%>\% group_by(sortkode, virkesgrp, sortiment, virkeskat, kategoritekst, treslag, grovsortiment) \%>\% summarise(vm = sum(totalvolum)) \%>\% arrange(sortkode, virkesgrp, virkeskat) \%>\%   print(n = 40)
}
