% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sortimentpriser_fylke_ldep}
\alias{sortimentpriser_fylke_ldep}
\title{sortimentpriser per fylke, treslag, sortiment, år fra 2014}
\format{
tibble med om lag 1300 obs og 10 variabler
\describe{
\item{reg_n202x}{Regionnavn pr dagens årstall chr "Agder", "Agder" }
\item{reg_k202x}{Regionkode pr dagens årstall chr "42", "42" }
\item{aar}{årstall; num  2014 2014 2014 ...}
\item{sortkode}{Sortimentkode : chr "01" "01" "01" ...}
\item{sortiment}{Sortimentnavn: chr "Sagtømmer" "Sagtømmer" ... }
\item{virkesgrp}{Virkesgruppe; chr "1-Gran" "1-Gran" "2-Furu" "3-Lauv" "5-Ved" ... }
\item{totalvolum}{m3; num 3440 14215 ...}
\item{totalverdi}{kr; num 1599470 6898324 703000 ...}
\item{m3pris}{kr pr m3; num 465 485 263 398 3 ...}
}
}
\usage{
sortimentpriser_fylke_ldep
}
\description{
Data aggregert fra Landbruksdirektoratets hogststatistikk
}
\keyword{datasets}
