% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sortimentpriser_kmn}
\alias{sortimentpriser_kmn}
\title{Sortimentpriser fordelt på kommune og år fra 1996}
\format{
data.frame med om lag 9542 obs og 7 variabler:
\describe{
\item{region}{Navn paa kommunen (paa hogsttidspunkt); chr "Hele landet" "Halden" ...}
\item{kategoritekst}{Sortimentnavn; chr "Gran spesialtoemmer" "Furu sekunda sagtoemmer" ...}
\item{statistikkvariabel}{ ; chr "Gjennomsnittspris" "Gjennomsnittspris"}
\item{år}{årstall ; chr "2006", "2007"}
\item{pris}{nok; int NA, 637, 465, ...}
\item{region_kode}{Kommunenummer paa registreringstidspkt: chr "0" "0101" "0102"}
\item{treslag}{chr  "Gran"   "Furu"   "Lauv"   "Ukjent"}
\item{virkeskategori}{virkeskategori chr "1110" "1141" "1143" "1148" "1160" "1410" "1490" "2110" "2141" "2143" "2148" "2160" "2410" "2490" "3120" "3400" "3800" "1800"}
\item{sortimentgruppe}{sortimentgruppe chr "tømmer" "sams" "massevirke" "annet"}
\item{aar}{Aarstall; int "1996" "1997" ...}
\item{reg_n2020}{Regionnavn (kommunenavn) pr 2020 chr "Hele landet" "Halden" "Moss" "Sarpsborg" "Fredrikstad"}
\item{reg_k2020}{Regionkode (kommunekode) pr 2020}
}
}
\source{
\url{https://www.ssb.no/statbank/table/12750}
}
\usage{
sortimentpriser_kmn
}
\description{
Data fra SSB, tabell 12750: Bruttoverdi. A (K) 2006 - dd
}
\keyword{datasets}
