% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R
\name{t03895}
\alias{t03895}
\title{Skogsavvirkning volum t03895
Hogststatistikk for tømmer SSB tabell 03895
1996 - 2019}
\usage{
t03895(region_level = "fylke")
}
\arguments{
\item{region_level}{regionnivå; velg ett av landet, fylke or kommune}
}
\value{
en tibble
}
\description{
Tabellen gir avvirkningsvolum for salg, etter sortiment, kommune, år.
Volum er avregningsvolum (m3pris)
}
\examples{
t03895()
}
