% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R
\name{t06216}
\alias{t06216}
\title{Skogsavvirkning priser t06216
prisstatistikk for rundvirke SSB tabell 06216}
\usage{
t06216()
}
\value{
en tibble med hele datasetet.
}
\description{
Tabellen gir snittpris per sortiment per fylke, fra 1996 til 2017.
Virke blir også klassifisert etter treslag og sortimentgrupper (skur, massevirke, sams, annet)
https://www.ssb.no/statbank/list/skogav#'
}
\examples{
t06216()
}
