% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t08801_ImpExp_Trevarer.R
\name{t08801tre}
\alias{t08801tre}
\title{Import Export wood assortments, pr year.}
\source{
\url{https://www.ssb.no/statbank/table/08801/}
}
\usage{
t08801tre()
}
\value{
a tibble with the entire dataset
}
\description{
Import and export 1996 - now.
Not sure if energywood assortments and wood is included
}
\examples{
trestat <-  t08801tre()
}
