% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R
\name{t12750}
\alias{t12750}
\title{Skogsavvirkning priser t12750
prisstatistikk for virke fra SSB tabell 12750}
\usage{
t12750()
}
\value{
en tibble med hele datasetet.
}
\description{
Tabellen gir snittpris per sortiment per fylke, fra 2006 til 2019.
Virke blir klassifisert per treslag og sortimentgrupper (skur, massevirke, sams, annet)
https://www.ssb.no/statbank/list/skogav
}
\examples{
t12750()
}
