% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virkesverdi_flk}
\alias{virkesverdi_flk}
\title{Brutto virkesverdi fordelt p책 fylke og 책r fra 1996}
\format{
data.frame med om lag 9542 obs og 7 variabler:
\describe{
\item{region}{Navn paa kommunen (paa hogsttidspunkt); chr "Hele landet" "Halden" ...}
\item{statistikkvariabel}{ ; chr "Bruttoverdi" "Bruttoverdi"}
\item{책r}{Aarstall; chr "1996" "1997" ...}
\item{bruttoverdi}{Brutto verdi av solgt industrivirke NOK: int 2709220 2877766 2608427 ...  }
\item{region_kode}{Regionkode <chr> "0101", "0101", }
\item{Tid}{samme som 책rstall, <chr> "1996", "1997", "1998",}
\item{aar}{Aarstall; int "1996" "1997" ...}
\item{reg_n2020}{Regionnavn (kommunenavn) pr 2020 chr "Hele landet" "Halden" "Moss" "Sarpsborg" "Fredrikstad"}
\item{reg_k2020}{Regionkode (kommunekode) pr 2020}
}
}
\source{
\url{https://www.ssb.no/statbank/table/03794}
}
\usage{
virkesverdi_flk
}
\description{
Data fra SSB, tabell 03794: Bruttoverdi. Avvirkning for salg (1 000 kr) (K) 1996 - dd
}
\keyword{datasets}
