% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R
\name{t03794}
\alias{t03794}
\title{Skogsavvirkning bruttoverdi t03794}
\source{
\url{https://www.ssb.no/statbank/table/03794/}
}
\usage{
t03794(region_level = "fylke")
}
\arguments{
\item{region_level}{Select type oneof: "fylke", "kommune"}
}
\value{
en tibble med hele datasetet.
}
\description{
bruttoverdi av tømmer (alt virke), pr år, fylke / kommune,
}
\details{
totalverdi av tømmer solgt per år og geografisk enhet, fra 1996 - dd.
Litt usikker om energivirkesortimenter og ved er med.
https://www.ssb.no/statbank/list/skogav
}
\examples{
 t03794()
}
