% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t03794.R
\name{t03794}
\alias{t03794}
\title{Skogsavvirkning bruttoverdi t03794}
\source{
\url{https://www.ssb.no/statbank/table/03794/}
}
\usage{
t03794(region_level = "fylke")
}
\arguments{
\item{region_level}{Select type oneof: "fylke", "kommune"}
}
\value{
a tibble with the entire dataset
}
\description{
gross value of roundwood, pr year, county / municipality,
}
\details{
total value of all roundwood sold per year and geographical unit, 1996 - now.
Not sure if energywood assortments and wood is included
https://www.ssb.no/statbank/list/skogav
}
\examples{
 t03794()
}
