% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t03895.R
\name{t03895}
\alias{t03895}
\title{Skogsavvirkning volum t03895
Hogststatistikk for rundvirke SSB tabell 03895
1996 - dd}
\usage{
t03895(region_level = "fylker2020", periode = TRUE)
}
\arguments{
\item{region_level}{regionnivaa; velg ett av landet, fylke or kommune}
}
\value{
en tibble
}
\description{
Tabellen gir avvirkningsvolum for salg, etter sortiment, kommune, year.
Volum er avregningsvolum (m3pris)
}
\examples{
volumstat <- t03895()
volumstat2 <- t03895(region_level = "fylke")
volumstat3 <- t03895(region_level = "kommune")
volumstat4 <- t03895(region_level = "fylke", periode = c("2021"))
str(volumstat)
}
