% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R
\name{t03895}
\alias{t03895}
\title{Skogsavvirkning volum t03895
Hogststatistikk for rundvirke SSB tabell 03895
1996 - dd}
\usage{
t03895(region_level = "fylke")
}
\arguments{
\item{region_level}{regionnivaa; velg ett av landet, fylke or kommune}
}
\value{
en tibble
}
\description{
Tabellen gir avvirkningsvolum for salg, etter sortiment, kommune, year.
Volum er avregningsvolum (m3pris)
}
\examples{
t03895() \%>\% dplyr::glimpse()
}
