% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t03895.R
\name{t03895}
\alias{t03895}
\title{Skogsavvirkning volum t03895
Hogststatistikk for rundvirke SSB tabell 03895
1996 - dd}
\usage{
t03895(region_level = c("fylker", "kommuner")[1])
}
\arguments{
\item{region_level}{regionnivaa; velg ett av landet, fylke or kommune}
}
\value{
en tibble
}
\description{
Tabellen gir avvirkningsvolum for salg, etter sortiment, kommune, år.
Volum er avregningsvolum (m3pris)
}
\examples{
volumstat <- t03895()
volumstat2 <- t03895(region_level = "fylke")
volumstat2 \%>\% filter(region == "Trøndelag" ) \%>\% group_by(region, aar) \%>\% summarise(m3 = sum(m3, na.rm =TRUE)) \%>\%  head()
volumstat3 <- t03895(region_level = "kommune")
unique(volumstat3$region)
volumstat3 \%>\% filter(region == "Steinkjer") \%>\% group_by(region, aar) \%>\% summarise(m3 = sum(m3, na.rm = TRUE)) \%>\%  tail()
}
