% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t08801_ImpExp_Trevarer.R
\name{t08801tre}
\alias{t08801tre}
\title{Import Export wood assortments, pr year.}
\source{
\url{https://www.ssb.no/statbank/table/08801/}
}
\usage{
t08801tre()
}
\value{
a tibble with the entire dataset
}
\description{
Import and export 1996 - now.
Not sure if energywood assortments and wood is included
}
\examples{
 dt <- t08801tre()
 str(dt)
 dt \%>\% mutate(kr_pr_m3 = ifelse(!is.na(m3) & m3>0, kr / m3, NA_real_)) \%>\%
  filter(!(stringr::str_detect(.data$grovsort, "440") )) \%>\%
   dplyr::filter(.data$kr_pr_m3 > 2000) \%>\%
  arrange(desc(kr_pr_m3)) \%>\% select(Tid, Varekoder, ImpEks, grovsort, assortment, m3, kg, kr, kr_pr_m3, norsk) \%>\% str()
}
