% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_skogsavvirkning.R,
%   R/t12750_pris_sort_flk_2006_23.R
\name{t12750}
\alias{t12750}
\title{Skogsavvirkning priser t12750
prisstatistikk for virke fra SSB tabell 12750}
\usage{
t12750()

t12750()
}
\value{
en tibble med hele datasetet.

en tibble med hele datasetet.
}
\description{
Tabellen gir snittpris per sortiment per fylke, fra 2006 til dd.
Virke blir klassifisert per treslag og sortimentgrupper (skur, massevirke, sams, annet)
https://www.ssb.no/statbank/list/skogav

Tabellen gir snittpris per sortiment per fylke, fra 2006 til dd.
Virke blir klassifisert per treslag og sortimentgrupper (skur, massevirke, sams, annet)
https://www.ssb.no/statbank/list/skogav
}
\examples{
prisstat <- t12750()
prisstat <- t12750()

}
